completed metronome

+++ src/com/jimjeb/bpmetronome/model/metronome/CurrentBeatObserver.java
@@ -2,5 +2,5 @@ package com.jimjeb.bpmetronome.model.metronome;
 
 public interface CurrentBeatObserver
 {
-    public abstract void update(int currentBeat);
+    public abstract void update(double currentBeat);
 }
